Refactor (CS majorify) the code a little bit.
I kid, but honestly, the code was getting a bit grody for me,
so I'm glad I did it now rather than later.

Instead of patches being simply a single function, there is now
a abstract "Patch" class which can be used to write patches.

I think this is a bit cleaner. Also, all of the SCE includes have been
kicked out of main.cpp, which is sorta nice.