Refactor patch accessors
Rather than having to define a bunch of singleton functions,
patches now can be registered with an ID, and you can just
use GetPatchById(...) to get the patch. It's cleaner, a little slower,
but it works really well, and is quite a bit more flexible.

It did make the binary a little fatter, though, which is a shame.
But, you win some, you lose some, and I'm sure i can thin out
stuff if it really is such a big problem.