Add ERL build support to build system, fix up ERL...
Adds support to tools.mk to build ERL files.
Fixes up and adds the sample ERL as a build product (not as a main part
of the build yet.).

Moves utils.cpp to src/debugout.cpp. I might move this file back, and
use __builtin_* routines & stuff (as well as a define to switch from
Newlib printf() to bx::printf()) so it can be a part of libutils again.

Messes with the ABI a tiny bit, however it's just type prettiness.