HostFS: COMPLETELY DETACH THE CD
FINALLY. I've also decided to ditch hardcoding the path,
and now use the fact PCSX2 sets the cwd of HostFS to where the ELF is
when ran.

This finally means the CD can be fully ditched.

Also contains bigless world support. Extract all the world BIGs and move
the files to where they should be (data/models).