Remove compile-time SSX 3 support
This is going to be replaced with a better runtime system,
which will scan and automatically probe the proper game version
of a game root ElfLdr is placed into.

This commit also updates the README.md to provide a game support table
and a setup guide for the HostFS patch.