Move to CMake
This is probably the craziest thing I've ever done.

Removes the ENTIRE tools.mk/Make build system and replaces it with a
sane CMake buildsystem.

The reasoning for this was simple:
  - Make is not scalable for large codebases, which elfldr is
  - Make is not friendly to other platforms (e.g: Mac or Windows)
  - Make just lowkey sucks in 2022.

This should also allow CLion integration to be a LOT better,
only needing probably env vars or some toolchain stuff.

The sample ERL still needs to be ported (hold on :V) to CMake