Runtime: Create library from utils sources
This makes a new library available to use in all elfldr projects,
whether ERL or ELF, which provides sane building blocks for C++ development
without the usage of the standard C++ library.

This commit marks provisions for also providing the C runtime inside
this library, but does not implement it (to make this a bit less tedious
on me).

Finally, this commit also completes (oops!) the de-C++ification of elfldr
includes, so that we now can replace the C includes with what we pay for.
Pretty cool.