*: Some cleanup
Utils now has a (currently untested, and WIP) API for function hooking,
which should be better than what I've been planning with the ERL abi.

It also has an internal header for MIPS instruction encoding (more
specifically, MIPS-I, which is valid for the PS2 as well)

On the ERLABI: Some adjustments will be made.

CMake buildsystem now has a Playstation2 platform and the main
CMakeLists.txt was edited to test for the CMAKE_SYSTEM_NAME we set.

The Sample erl was ported to cmake but doesn't build all the way (pesky
linker flags...I'm considering a total hack which probably
won't be worth it)

Really most of this is buildsystem and code cleanup. Stuff I needed
to do a while ago :P