mlstd: Implement string SSO
SSO (small string optimization) is a optimization where the string object can either use a pointer to memory on the heap or a small buffer part of the object for storage (for sufficiently small strings).

String is now ~16 bytes (depending on sizeof(Alloc)); which allows for about that long string size. Note that wider strings may only be able to store fewer characters.